Val fields can be barred, Link/Swap cannot have speculative values
When a `val` or `spec`-field is barred in a pristine type, it means that
it contains some speculative value. If `y : Foo | g`, `CAT(x.f, z, y)`
is not allowed, but `CAT(x.f, y.g, y)` is, as `y.g` acts as a witness
that the speculation is valid.