Support for thread mode
Capabilities with the thread mode may not leave the active object that
created it. This is checked when calling methods on active objects; if
an argument is thread or the return type is thread, the target must
either be this, or itself be thread. This also means that fields with
the thread mode can only be declared in traits that are manifestly
thread.