Experimental support for spine mode
The `spine` mode is a subordinate mode that must also be treated
linearly. This allows it to be used as the target of `CAT`s. The reason
for not using `linear + subord` is that this type is a nested dominating
aggregate whose subordinate fields (e.g. `Node.next`) may not be
accessed from the outside.