Experimental splitting support
A let expression can now have several variables *per expression*, and
each variable can optionally have a type annotation:
```
let i, f : float = 42;
print i; -- 42
print f; -- 42.000000
```
If the types of the variables are linear, these are checked to be in
conjunction in the type of the right hand side:
```
let p = new Person();
let a : Aged, n : Named = consume p;
```
If the types are not linear any splitting is allowed. Not giving a type
to a variable defaults to the type of the right hand side (meaning any
splitting will fail).