Fix #452, consider safe composite types safe
This commit considers `Maybe T`, `Stream T`, `Fut T` and `Par T` as safe
types iff `T` is a safe type. Tuples are safe if all the constituent
types are safe. Ranges are also safe.

Arrays are not safe since they can be written to. In the future one
could consider a type like `[val T]` to be safe (an array that can only
be read from). I'll leave that design up to @glundi.